requirement.txt
langchain_community
langchain_huggingface
gradio
faiss-cpu
sentence-transformers
transformers
streamlit>=1.28.0
faiss-cpu>=1.7.0
google-generativeai>=0.3.0
python-dotenv>=0.19.0
numpy>=1.21.0
sentence-transformers>=2.2.0
pypdf2>=1.27.0
python-docx>=0.8.11